move tests to sqlite/test/vector.test.js
